libevent: fix @hellais and @sarathms linux builds
The configure script of libevent is such that, if we don't have
either `pkg-tool` or `libssl-dev` installed, it does not build
`libevent_openssl`, which is not what we want.

The simplest fix it to let libevent believe it succeded to
find OpenSSL using `pkg-config`. A more comprehensive fix
would be to improve the `libevent` configure script to allow
for override the location when OpenSSL is installed.