add framework-ios: script to generate iOS frameworks
I've figured out that it's much better doing it here rather than
having a separate repository only meant to do that.

Compared to previous attempts, this one is much better in that it
recognizes whether a specific framework has machine dependent headers
or not. If they are not machine dependent, headers are installed all
inside of the `Headers` directory. Otherwise, we install them using
the target architecture as the prefix.

Allows, downstream, to include headers of dependencies directly
into the iOS build. Previously, we could only link to dependencies
but their headers were actually not available.